third_party/blink/web_tests/platform/generic/fast/forms/radio/interactive-validation-required-radio-expected.txt

There was a bug of validation of a required radio button. This test confirms the bug has been fixed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS The form should be submitted.
PASS location.search.indexOf("terms=on") != -1 is true
TEST COMPLETE
  
